refactor: create module for build subcommand
